Assume video resolution in a 16:9 matte canvas
This is for a more accurate resolution check in the ffmpeg.auto config. E.g. 1920x1036 (1.85:1) video would be considered `1036` resolution, which isn't going to match the ffmpeg.auto config values, unless you specifically define 1036 (aint nobody got time for that).

With this change that same video would be considered letterboxed 1920x1080 (1.85 in a 16:9 matted canvas), so `1080` which would match the config definitions.